Revert "fix: use matchedData() instead of req.query for validated params"
This reverts commit ce5f9daa85263caf5c478df630a25ae546759c68.

+++ src/models/twilioMessages.js
@@ -7,8 +7,8 @@ export const twilioMessageValidation = [
 ]
 
 export const twilioSearchValidation = [
-    query('since').optional().isInt().toInt(),
-    query('until').optional().isInt().toInt(),
+    query('since').optional().isInt(),
+    query('until').optional().isInt(),
     query('from').optional().isString(),
     query('to').optional().isString()
 ]
